@@ -1,5 +1,5 @@
 // ASN.1 RFC definitions matcher
-// Copyright (c) 2023-2023 Lapo Luchini <lapo@lapo.it>
+// Copyright (c) 2023-2024 Lapo Luchini <lapo@lapo.it>
 
 // Permission to use, copy, modify, and/or distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -13,13 +13,7 @@
 // ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 // OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-(typeof define != 'undefined' ? define : function (factory) { 'use strict';
-    if (typeof module == 'object') module.exports = factory(function (name) { return require(name); });
-    else window.defs = factory(function (name) { return window[name.substring(2)]; });
-})(function (require) {
-'use strict';
-
-const rfc = require('./rfcdef');
+import { rfcdef } from './rfcdef.js';
 
 function translate(def, tn, stats) {
     if (def?.type == 'tag' && !def.explicit)
@@ -56,14 +50,14 @@ function firstUpper(s) {
     return s[0].toUpperCase() + s.slice(1);
 }
 
-class Defs {
+export class Defs {
 
     static moduleAndType(mod, name) {
         return Object.assign({ module: { oid: mod.oid, name: mod.name, source: mod.source } }, mod.types[name]);
     }
 
     static searchType(name) {
-        for (const mod of Object.values(rfc))
+        for (const mod of Object.values(rfcdef))
             if (name in mod.types) {
                 // console.log(name + ' found in ' + r.name);
                 // return r.types[name];
@@ -122,7 +116,7 @@ class Defs {
 
 }
 
-Defs.RFC = rfc;
+Defs.RFC = rfcdef;
 
 Defs.commonTypes = [
     [ 'X.509 certificate', '1.3.6.1.5.5.7.0.18', 'Certificate' ], 
@@ -133,8 +127,4 @@ Defs.commonTypes = [
     [ 'PKCS#8 private key', '1.2.840.113549.1.8.1.1', 'PrivateKeyInfo' ],
     [ 'PKCS#10 certification request', '1.2.840.113549.1.10.1.1', 'CertificationRequest' ],
     [ 'CMP PKI Message', '1.3.6.1.5.5.7.0.16', 'PKIMessage' ],
-].map(arr => ({ description: arr[0], ...Defs.moduleAndType(rfc[arr[1]], arr[2]) }));
-
-return Defs;
-
-});
+].map(arr => ({ description: arr[0], ...Defs.moduleAndType(rfcdef[arr[1]], arr[2]) }));
